+++ pyproject.toml
@@ -5,18 +5,25 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
+    "alembic>=1.13.3",
+    "asyncpg>=0.29.0",
     "beautifulsoup4>=4.12.3",
-    "fastapi>=0.115.0",
+    "fastapi[standard]>=0.115.0",
     "playwright>=1.47.0",
+    "pre-commit>=3.8.0",
     "python-decouple>=3.8",
     "requests>=2.32.3",
+    "sqlalchemy>=2.0.34",
     "sqlmodel>=0.0.22",
-    "uvicorn>=0.30.6",
 ]
+
 [tool.uv]
 dev-dependencies = [
+    "pgcli>=4.1.0",
+    "psycopg2-binary>=2.9.9",
     "ruff>=0.6.6",
 ]
+
 
 [tool.ruff]
 line-length = 120
@@ -35,32 +42,8 @@ exclude = [
     "site-packages",
     "venv",
 ]
-
-[tool.ruff.lint]
-extend-select = ["C4", "SIM", "TCH"]
-ignore =["D205","D401","D211","D212", "D415","D400","D100"]
-select = [
-    "B",  # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "D",  # pydocstyle
-    "E",  # Error
-    "F",  # pyflakes
-    "I",  # isort
-    "N",  # pep8-naming
-    "PGH",  # pygrep-hooks
-    "PTH", # flake8-use-pathlib
-    "Q",  # flake8-quotes
-    "S",  # bandit
-    "SIM",  # flake8-simplify
-    "TRY",  # tryceratops
-    "UP",  # pyupgrade
-    "W",  # Warning
-    "YTT", # flake8-2020
-]
-
 [tool.ruff.format]
 quote-style = "double"
 indent-style = "tab"
 docstring-code-format = true
 skip-magic-trailing-comma = false
-
