+++ src/constants.py
@@ -1,6 +1,5 @@
 """WEBSCRAPING PAGES CONSTANTS"""
+
 SOURCE_ONE = "https://do512.com"
 SOURCE_TWO = "https://heyaustin.com/austin-events/"
 SOURCE_THREE = "https://austin.culturemap.com/events/?tags=20240920"
-SOURCE_FOUR = "https://www.austintexas.org/event/"
-
